done the expenditure page

+++ SandTetris/Entities/SalaryDetailSummary.cs
@@ -6,6 +6,6 @@ public class SalaryDetailSummary
     public string DepartmentName { get; set; }
     public int Month { get; set; }
     public int Year { get; set; }
-    public decimal TotalSpent { get; set; }
+    public int TotalSpent { get; set; }
     public string DateString => $"{Month}/{Year}";
 }
